refactor(admin): getRequestBody 세분화
