+++ apps/admin/src/services/fair/mutations.ts
@@ -1,5 +1,5 @@
 import { useMutation } from '@tanstack/react-query';
-import type { FairApiRequestBody } from '@/utils/functions/getRequestBody';
+import type {FairApiRequestBody} from '@/components/fair/FairForm/fair.hooks';
 import { postFairReq } from '@/services/fair/api';
 import { useApiError } from '@/hooks';
 import { toast } from 'react-toastify';
@@ -9,7 +9,9 @@ export const useCreateFairMutation = () => {
   return useMutation({
     mutationFn: (data: FairApiRequestBody) => postFairReq(data),
     onSuccess: () => {
-      toast.success('입학 설명회 일정이 만들어졌습니다.', { type: 'success' });
+      toast.success('입학 설명회 일정이 만들어졌습니다.', {
+        type: 'success' }
+      );
     },
     onError: handleError,
   });
