+++ apps/admin/src/utils/functions/convertToApiDateFormat.ts
@@ -1,0 +1,12 @@
+const convertToApiDateFormat = (dateStr: string | null): string => {
+  if (!dateStr) return '';
+  if (/^\d{4}-\d{2}-\d{2}$/.test(dateStr)) return dateStr;
+  if (/^\d{8}$/.test(dateStr)) {
+    const y = dateStr.slice(0, 4);
+    const m = dateStr.slice(4, 6);
+    const d = dateStr.slice(6, 8);
+    return `${y}-${m}-${d}`;
+  }
+  return '';
+};
+export default convertToApiDateFormat;
