--- apps/admin/src/utils/functions/formatFairDate.ts
@@ -1,50 +1,0 @@
-import { useFairFormStore } from '@/store/fair/fairType';
-import { formatFairRequestBody } from '@/utils/functions/getRequestBody';
-import formatDate from './formatDate';
-import type { FairFormInput } from '@/utils/functions/getRequestBody';
-
-const useFairForm = () => {
-  const [form, setForm] = useFairFormStore();
-
-  const handleChange = (key: keyof FairFormInput, value: string) => {
-    setForm((prev) => ({
-      ...prev,
-      [key]: value,
-    }));
-  };
-
-  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const { name, value } = e.target;
-    handleChange(name as keyof FairFormInput, value);
-  };
-
-  const handleDateChange = (value: string) => {
-    const newStart = value + form.start.slice(8);
-    setForm((prev) => ({ ...prev, start: newStart }));
-  };
-
-  const handleTimeChange = (value: string) => {
-    const newStart = form.start.slice(0, 8) + value;
-    setForm((prev) => ({ ...prev, start: newStart }));
-  };
-
-  const getRequestBody = () => {
-    return formatFairRequestBody({
-      ...form,
-      applicationStartDate: formatDate.toDashedDate(form.applicationStartDate ?? ''),
-      applicationEndDate: formatDate.toDashedDate(form.applicationEndDate ?? ''),
-    });
-  };
-
-  return {
-    form,
-    setForm,
-    handleChange,
-    handleInputChange,
-    handleDateChange,
-    handleTimeChange,
-    getRequestBody,
-  };
-};
-
-export default useFairForm;
