+++ apps/admin/src/utils/index.ts
@@ -2,4 +2,3 @@ export { default as resizeTextarea } from './functions/resizeTextarea';
 export { default as formatDate } from './functions/formatDate';
 export { default as formatPhoneNumber } from './functions/formatPhoneNumber';
 export { default as convertToResponsive } from './functions/convertToResponsive';
-export { default as formatFairDate } from './functions/formatFairDate';
